Add possibility to favorite subject from subject page
